Don't kill nonexisting yet instance (Fix #102)

+++ hyphe_frontend/app/views/network.js
@@ -290,8 +290,10 @@ angular.module('hyphe.networkController', [])
     }
 
     function killSigma(){
-      $scope.stopSpatialization()
-      $scope.sigmaInstance.kill()
+      if ($scope.sigmaInstance) {
+        $scope.stopSpatialization()
+        $scope.sigmaInstance.kill()
+      }
     }
 
     function buildNetwork(){
